+++ src/main/java/br/com/hitbox/interfaces/dto/response/kanban/KanbanCardResponse.java
@@ -4,6 +4,7 @@ import lombok.*;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Getter
 @Setter
@@ -24,7 +25,7 @@ public class KanbanCardResponse {
 
     private BigDecimal productionProgress;
 
-    private Long estimatedMinutes;
+    private BigDecimal estimatedMinutes;
 
     private Long actualMinutes;
 
@@ -48,6 +49,6 @@ public class KanbanCardResponse {
 
     private String clientName;
 
-    private Long quantity;
+    private BigDecimal quantity;
 
 }
